requirements.txt
aniso8601==9.0.1
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
dnspython==2.2.1
dominate==2.6.0
email-validator==1.2.0
filelock==3.6.0
Flask==2.1.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-Login==0.6.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
huggingface-hub==0.5.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
MarkupSafe==2.1.1
numpy==1.22.3
packaging==21.3
pyparsing==3.0.8
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
sacremoses==0.0.49
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.35
tokenizers==0.12.1
tqdm==4.64.0
transformers==4.18.0
typing_extensions==4.2.0
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.1.1
WTForms==3.0.1
